Finish the remaining functions for the new SDP representation. Untested.
Remove union find, not usefull finaly. The new architecture proseeds to solve the problem in 2 steps, each associated with a data structure. An algebraic/symbolic step (in momentmatrix.py for now) and the relaxation to a proper SDP, in semidefinite_porogram_repr.py. Translation between the two is done in problem.py. The second step is fed into solvers in solver.py, only cvxpy for now.